JVMCBC-655 UnsignedLEB128 outputs bytes in reversed order
Motivation
==========
The utility was generating incorrect encoded values with the byte
order swapped.

Switch the utility back to using longs instead of BigIntegers.
BigInteger arithmetic generates garbage, and we don't need to handle
arbitrarily large values.

(I think previously there was some confusion about whether the "128"
in LEB128 meant the values needed to be 128 bits wide. While the LEB128
format itself does not limit the magnitude, in practice the values
received from the server are no longer than 64 bits.)

Add some routines for reading and writing unsigned LEB128 values
directly from/to a Netty ByteBuf without an intermediate byte array.

Reviewed-on: http://review.couchbase.org/108132
Tested-by: Build Bot <build@couchbase.com>

+++ core-io/src/integrationTest/java/com/couchbase/client/core/util/CoreIntegrationTest.java
@@ -44,7 +44,7 @@ public class CoreIntegrationTest extends ClusterAwareIntegrationTest {
   /**
    * Calculates the default collection id.
    */
-  public static byte[] DEFAULT_COLLECTION_ID = UnsignedLEB128.encode(BigInteger.valueOf(0));
+  public static byte[] DEFAULT_COLLECTION_ID = UnsignedLEB128.encode(0);
 
   /**
    * Creates an environment builder that already has all the needed properties to bootstrap
